Extract logscan cache/path primitives into modules/logscan_cache.py (1/6) (#1312)
